Gameplay score update 2#
When the user presses the correct or wrong button the answer is stored as a bool value, the code then checks to see if a record currently exists for that user with the current card, if it doesn't exist it adds the record to the score table, if it does exist it updates the record.

Next will be redirecting the page to the next card as that functionalit is now broken with the link taking the user to score.inc.php.