Results page functionality
The first version of the results page has been created, having a simple design with hover effects. Links currently work, the retry button redirects the user to the category selected at the first card in the set, this may need to change when the random feature is implemented. The results, image and category details are static, they need to be programmed to display the users data.